[Feature request] Alphabetic subhintls
Could the subhint for windows from applications with multiple windows be mapped from numbers to letters? Similar to how hs.hints and vimium does it? Ideally, it'd be a configurable option and set of characters like hs.hints//vimium. I'm a big fan of the visual expose, but wish it was even more keyboard home row friendly!

I'm sadly not familiar with Lua and could only get a vague sense of the `normalize` function and the data structure it seems to be recursively mutating so the proposed "changes" are totally just pseudocode.

Anyways, much thanks